In plugin Move Subtitles, move selected or move all subsequent based on selection
Up until now, there was a user configuration whether only selected
subtitles are to be moved, or the first selected and all subsequent
subtitles are to be moved. This is getting rid of this setting and
instead decides based on selection. When one subtitle is selected, this
subtitled and all subsequent subtitles are moved. When more subtitles
are selected, only those are moved.