In Configure Keyboard Shortcuts, sort them alphabetically and show dynamic ones first
The keyboard shortcuts were a mess without any order,
it was difficult to find anything there.
Now they are sorted alphabetically and dynamic ones
(open documents, audio tracks and views) are shown first.
They are also separated by a separator.
The code is a bit repetetive, but I think it is more readable this way
and the sorting is a bit different in each case anyway.